Foundation: add some type conversion extensions for Windows
The file attribute flags are used extensively in the file manipulation
code.  Explicitly converting to the value to `DWORD` adds no value and
clutters the code.  Provide an internal overload to translate the type
appropriately.

+++ Sources/Foundation/FileManager.swift
@@ -431,8 +431,8 @@ open class FileManager : NSObject {
                     }
 
                     let hiddenAttrs = isHidden
-                        ? attrs | DWORD(FILE_ATTRIBUTE_HIDDEN)
-                        : attrs & ~DWORD(FILE_ATTRIBUTE_HIDDEN)
+                        ? attrs | FILE_ATTRIBUTE_HIDDEN
+                        : attrs & ~FILE_ATTRIBUTE_HIDDEN
                     guard SetFileAttributesW(fsRep, hiddenAttrs) else {
                       throw _NSErrorWithWindowsError(GetLastError(), reading: false, paths: [path])
                     }
@@ -604,7 +604,7 @@ open class FileManager : NSObject {
 
 #if os(Windows)
         let attrs = attributes.dwFileAttributes
-        result[._hidden] = attrs & DWORD(FILE_ATTRIBUTE_HIDDEN) != 0
+        result[._hidden] = attrs & FILE_ATTRIBUTE_HIDDEN != 0
 #endif
         result[.ownerAccountID] = NSNumber(value: UInt64(s.st_uid))
         result[.groupOwnerAccountID] = NSNumber(value: UInt64(s.st_gid))
@@ -1295,9 +1295,9 @@ public struct FileAttributeType : RawRepresentable, Equatable, Hashable {
 
 #if os(Windows)
     internal init(attributes: WIN32_FILE_ATTRIBUTE_DATA, atPath path: String) {
-        if attributes.dwFileAttributes & DWORD(FILE_ATTRIBUTE_DEVICE) == DWORD(FILE_ATTRIBUTE_DEVICE) {
+        if attributes.dwFileAttributes & FILE_ATTRIBUTE_DEVICE == FILE_ATTRIBUTE_DEVICE {
             self = .typeCharacterSpecial
-        } else if attributes.dwFileAttributes & DWORD(FILE_ATTRIBUTE_REPARSE_POINT) == DWORD(FILE_ATTRIBUTE_REPARSE_POINT) {
+        } else if attributes.dwFileAttributes & FILE_ATTRIBUTE_REPARSE_POINT == FILE_ATTRIBUTE_REPARSE_POINT {
             // A reparse point may or may not actually be a symbolic link, we need to read the reparse tag
             let handle: HANDLE = (try? FileManager.default._fileSystemRepresentation(withPath: path) {
               CreateFileW($0, /*dwDesiredAccess=*/DWORD(0),
@@ -1318,7 +1318,7 @@ public struct FileAttributeType : RawRepresentable, Equatable, Hashable {
                 return
             }
             self = tagInfo.ReparseTag == IO_REPARSE_TAG_SYMLINK ? .typeSymbolicLink : .typeRegular
-        } else if attributes.dwFileAttributes & DWORD(FILE_ATTRIBUTE_DIRECTORY) == DWORD(FILE_ATTRIBUTE_DIRECTORY) {
+        } else if attributes.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY == FILE_ATTRIBUTE_DIRECTORY {
             // Note: Since Windows marks directory symlinks as both
             // directories and reparse points, having this after the
             // reparse point check implicitly encodes Windows
